chore: Update {{name.snakeCase()}} repository
This commit updates the {{name.snakeCase()}} repository by modifying the `read` and `refresh` methods in the `{{name.pascalCase()}}Repository` class. The changes include adding optional parameters for pagination, `page` and `limit`, to both methods. This allows for more flexible data retrieval and refreshing in the repository.

Additionally, this commit includes updates to the `reviewDependencies` function in the `review.dart` file. It registers the `UpdateReviewBloc` and `UpdateReviewUseCase` classes as dependencies, enabling the use of the `UpdateReviewUseCase` in the application.

Lastly, the `review.dart` file is modified to export the `UpdateBloc` class, which is used for updating reviews in the application.